memory/service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.26.3
requests==2.31.0
